chore: use http handler
