+++ services/greet/hello_handler/http_handler.go
@@ -1,0 +1,5 @@
+package hello_handler
+
+import "net/http"
+
+var HTTPHandler http.Handler = http.HandlerFunc(Handler)
