+++ services/notify/consumer_handler/handler.go
@@ -8,12 +8,9 @@ import (
 	mongolib "github.com/web-notify/api/monorepo/libs/db/mongo_lib"
 	webpush "github.com/web-notify/api/monorepo/libs/notifications/web_push"
 	qmodels "github.com/web-notify/api/monorepo/libs/queue/models"
-	appconfig "github.com/web-notify/api/monorepo/libs/store/app_config"
-	"github.com/web-notify/api/monorepo/libs/store/vault"
 
 	// appConfig "github.com/web-notify/api/monorepo/libs/store/app_config"
 
-	"github.com/web-notify/api/monorepo/libs/utils/config"
 	"github.com/web-notify/api/monorepo/libs/utils/formats"
 	"github.com/web-notify/api/monorepo/services/notify/lib"
 	"github.com/web-notify/api/monorepo/services/notify/models"
@@ -74,32 +71,4 @@ func handler(
 		http.Error(rw, err.Error(), http.StatusInternalServerError)
 		return
 	}
-}
-
-func Handler(rw http.ResponseWriter, req *http.Request) {
-	var stage config.STAGE = config.Stage()
-
-	// Get the VAPID keys
-	vaultService := vault.NewVaultService("https://kv-notify-secrets-stg.vault.azure.net")
-	appConfigService := appconfig.NewAppConfig("e53c986e-fa42-4065-bcef-9a5ae182d65a", "rg-webnotify-stg", "appcs-webnotify-stg")
-
-	vapidConf, err := lib.FetchVapidConfig(vaultService, appConfigService)
-	if err != nil {
-		http.Error(rw, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	formats.Trace(bson.M{
-		"vapidPublicKey":   vapidConf.PublicKey,
-		"vapidSenderEmail": vapidConf.Email,
-	})
-
-	webpushService := webpush.NewWebPush(
-		"olKYvd22Tt-OsP5-X25jxFfDAr4hlWFX6eeUX3i_D7I",
-		"BPlL5OTZwtW-0-4pQXmobTgX6URszc9-UKoTTvpvInhUlPHorlDM8y04J-rrErlQXMVH7_Us983mNmmwsb-z53U",
-		"eugenetham1994@gmail.com",
-	)
-	mongoService := mongolib.NewMongoService("subscriberDB", config.ENV_VARS[stage].MONGO_DB_CONNECTION_STRING)
-
-	// Dependency injection
-	handler(webpushService, mongoService, rw, req)
 }
