+++ services/notify/producer_handler/handler.go
@@ -1,12 +1,10 @@
 package producer_handler
 
 import (
-	"context"
 	"encoding/json"
 	"net/http"
 
 	qlib "github.com/web-notify/api/monorepo/libs/queue"
-	"github.com/web-notify/api/monorepo/libs/utils/config"
 	"github.com/web-notify/api/monorepo/libs/utils/formats"
 	"github.com/web-notify/api/monorepo/services/notify/models"
 )
@@ -51,20 +49,3 @@ func handler(qService qlib.QueueServicer, rw http.ResponseWriter, request *http.
 		return
 	}
 }
-
-// Dependency injection
-func Handler(response http.ResponseWriter, request *http.Request) {
-	queueName := "my-queue"
-	var stage config.STAGE = config.Stage()
-	queueAccountName := config.ENV_VARS[stage].QUEUE_ACCOUNT_NAME
-
-	qBaseUrl := config.QueueBaseURL(stage, queueAccountName)
-	var qService qlib.QueueServicer = qlib.NewQueueService(context.Background(), queueName, qBaseUrl, queueAccountName, config.ENV_VARS[stage].QUEUE_ACCOUNT_KEY)
-
-	handler(qService, response, request)
-}
-
-// var vaultService = vault.NewVaultService("https://kv-notify-secrets-stg.vault.azure.net")
-
-// Apply middleware
-var HTTPHandler http.Handler = http.HandlerFunc(Handler)
